完成 ColumnSeries

+++ UtilZ.Dotnet2020/UtilZ.DotnetCore.WindowEx/WPF/Controls/Chart/Axis/AxisConstant.cs
@@ -7,7 +7,7 @@ namespace UtilZ.DotnetCore.WindowEx.WPF.Controls
     public class AxisConstant
     {
         public const double ZERO_D = 0d;
-        public const double ZERO_I = 0d;
+        public const int ZERO_I = 0;
         public const double LABEL_TEXT_INTERVAL = 5d;
         public const int DEFAULT_STEP_SIZE = 100;
         public const double AXIS_DEFAULT_SIZE = 30d;
